Update PROJECT-REPORT.Rmd
Added data of interest from TO_social to TO_crime by neighbourhood so that it is mappable and comparable with the crime data.

Data added = population density, number of low income individuals, number of visible minority individuals, and population aged 15-29

Also made basic maps for pop density and % of individuals in each of the other 3 categories.